new commit

+++ cursova/src/main/java/com/example/modeling/components/CompDevice.java
@@ -41,7 +41,7 @@ public class CompDevice implements Component {
                 this.next.ifPresent(next -> next.process());
             }
         } else {
-            this.device.wait(time.doubleValue());
+            this.device.wait(time);
         }
     }
